# cloud-layer/cloud-tenant-registry/openapi.yaml
openapi: 3.0.0
info:
  title: Cloud Tenant Registry API
  version: 1.0.0
  description: Multi-tenant master data management service for FarmIQ
servers:
  - url: http://localhost:5121/api
    description: Local development server
paths:
...
  /v1/batches:
    get:
      summary: Get all batches for tenant
      tags: [Batches]
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
        - name: farmId
          in: query
          schema:
            type: string
        - name: barnId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of batches
    post:
      summary: Create a new batch
      tags: [Batches]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
      responses:
        '201':
          description: Batch created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateBatchRequest:
      type: object
      required:
        - species
        - farmId
        - barnId
      properties:
        species:
          type: string
        breedCode:
          type: string
          description: Canonical genetic-line code, e.g. Arbor Acres Plus
        sex:
          type: string
          enum: [as_hatched, male, female]
          description: Flock sex used by the XGBoost growth-standard lookup.
        initialHeadcount:
          type: integer
          minimum: 1
        farmId:
          type: string
          format: uuid
        barnId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, completed, cancelled]
...